some mobile changes + refactoring
