+++ src/WebFrontEnd/ClientApp/src/app/core/models/gamemodes.ts
