+++ src/WebFrontEnd/ClientApp/src/app/core/models/lobby.ts
