+++ src/WebFrontEnd/ClientApp/src/app/layout/nav-menu/nav-menu.component.scss
@@ -98,7 +98,7 @@
 }
 
 .navbar-homepage{ 
-  background: transparent;
+  background: var(--ornamental-pastelyellow);
   box-shadow: none;
   position: fixed;
   top:0;
