+++ src/WebFrontEnd/ClientApp/src/app/modules/home/components/gamemodes-section/gamemode-list/gamemode-list.component.scss
@@ -38,16 +38,10 @@ section {
   color:var(--black-secondary);
   font-weight: bold;
   font-size: 2em;
-  user-select: none;
-  opacity: 80%;
   transition: all 150ms cubic-bezier(0.25, 0.46, 0.45, 0.84);
 
   &:hover {
     transform: scale(1.1);
-  }
-
-  &:hover, &:focus {
-    opacity: 1;
   }
 
   /* Position the buttons */
@@ -89,5 +83,11 @@ section {
 
   .gamemode-list {
     padding: 0 1rem;
+  }
+}
+
+@media only screen and (max-width: 420px) {
+  .control {
+    padding: 0;
   }
 }
