+++ src/WebFrontEnd/ClientApp/src/app/modules/home/components/gamemodes-section/gamemode/gamemode.component.scss
@@ -54,3 +54,10 @@
         margin-left: 1rem;
     }
 }
+
+@media only screen and (max-width: 420px) {
+    .gamemode {
+        margin: 0;
+        padding: 0;
+    }
+}
