+++ src/WebFrontEnd/ClientApp/src/app/modules/home/components/hero-section/hero-section.component.html
@@ -3,9 +3,8 @@
     <div class="cast">
       <img [@simpleFadeAnimation]="'fastin'" class='hero-image' src="assets/svgs/hero-logo.svg"/>
     </div>
-    <h2 [@simpleFadeAnimation]="'slowin'" class="catchphrase">
-      Drawn Together.
-      Polled Apart.
-    </h2>
+    <div [@simpleFadeAnimation]="'slowin'" class="catchphrase">
+      <h2>Drawn Together.</h2><h2>Polled Apart.</h2>
+    </div>
   </div>
 </section>
