+++ src/WebFrontEnd/ClientApp/src/app/modules/home/components/hero-section/hero-section.component.scss
@@ -1,3 +1,10 @@
 section {
     padding: 0;
 }
+
+@media only screen and (max-width: 420px) {
+    .catchphrase {
+      flex-direction: column;
+      align-items: center;
+    }
+  }
