+++ src/WebFrontEnd/ClientApp/src/app/modules/lobby/components/gamemode-dialog/gamemode-dialog.component.html
@@ -14,7 +14,7 @@
             <!-- <input type="checkbox" id="bool-switch-{{i}}" checked="" [(ngModel)]="option.value" > -->
         </div>
     </div>
-    <div class="Error" *ngIf="error">
+    <div class="error" *ngIf="error">
         {{ error }}
     </div>
 </mat-dialog-content>
